Add CLI flag to disable DNS watcher.
I have this case: I'm connecting to my first VPN via Pritunl and then I'm connecting to the second VPN that gives me some routes and DNS servers. So, I need the ability to *not* reset the DNS setting on change. I think that this is the "developer" setting because I require to understand the consequences of this setting (i.e. you could lose the DNS server in your settings, etc.). So it could be used as an additional CLI flag in the `pritunl-service` utility.